+++ src/app/(content)/news/bulletins/_component/BulletinTable.tsx
@@ -7,7 +7,7 @@ import {
   getCoreRowModel,
   useReactTable
 } from '@tanstack/react-table';
-import Pagination from '@/components/layout/Pagination';
+import Pagination from '@/components/ui/Pagination/Pagination';
 import { ITEM_PER_PAGE } from '@/constants/bulletin';
 import type { BulletinType } from '@/types/common';
 import styles from './BulletinTable.module.scss';
