+++ src/components/ui/Pagination/Pagination.module.scss
